Restructure composables.
